requirements.txt
# Core dependencies for BMC AMI DevX Code Pipeline MCP Server
# FastMCP 2.x implementation with real MCP protocol support

# FastMCP - Real MCP server implementation
fastmcp>=2.12.2

# Data validation and settings management
pydantic>=2.0.0

# HTTP client for BMC AMI DevX Code Pipeline API
httpx>=0.28.0

# ASGI web framework for custom routes
starlette>=0.47.0

# System monitoring (optional)
psutil>=5.9.0

# Environment variable management
python-dotenv>=1.1.0

# OpenTelemetry observability dependencies (compatible versions)
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-httpx==0.55b1
opentelemetry-instrumentation-asyncio==0.55b1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-prometheus==0.55b1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-semantic-conventions==0.55b1
prometheus-client>=0.19.0

# Development and testing dependencies
pytest>=8.4.0
pytest-asyncio>=1.1.0
pytest-cov>=6.2.0
pytest-mock>=3.14.0

# Code quality tools
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
autoflake>=2.3.0
pre-commit>=3.8.0

# Coverage reporting
coverage>=7.10.0
